Move and rename LocalClientBucket to something more fitting.
