+++ .coveragerc
@@ -3,6 +3,7 @@ source = relstorage
 omit =
      relstorage/tests/*
      relstorage/adapters/tests/*
+     relstorage/cache/tests/*	 
      # Omit the main oracle adapter file since we don't test it at all
      relstorage/adapters/oracle.py
